requirements.txt
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
pinecone-client==3.0.0
openai==1.12.0
pypdf2==3.0.1
python-docx==1.1.0
tiktoken==0.5.2
